// src/protohax/fixtures/protohax_pb.ts
export enum Enum {
  /**
   * @generated from enum value: ENUM_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: ENUM_ONE = 1;
   */
  ONE = 1,

  /**
   * @generated from enum value: ENUM_TWO = 2;
   */
  TWO = 2,
}